[WIP] fix(virt_lines): invalidate win cursor position on redraw
If virtual lines are placed above or on the same line as the cursor, w_wcol may
be wrong as it's not adjusted for the added or removed lines.

This is only a problem if the buffer in which the lines were placed is the one
shown in the current window; if the window is not current, switching to it will
already update the cursor.