fix(lsp): do not assume `client.config.capabilities` exists
PR #23689 assumes `client.config.capabilities` exists, but this is only true
if that was passed to `vim.lsp.start{_client}`.

Fixes the issue in #23806 for 0.9.

This is only true prior to #23681, as after that PR, it's set to
`make_client_capabilities()` instead; that PR will not be backported, so just
fix it by checking nil instead.