fix(terminal): add various missing redraws
Problem: missing redraws when restoring saved cursorline/column, plus missing
statusline and mode redraws when not updating the screen in terminal mode.

Solution: schedule the redraws in a similar manner to other modes and remove
some now unnecessary redrawing logic.

Explicitly redraw if cursorline-related options change from entering terminal
mode. This fixes test failures in later commits.

Curiously, TextChangedT triggers based on must_redraw, which is... fun...?
Try to preserve its behaviour as much as we can here.

Interestingly, Vim's behaviour for TextChangedT seems different from a quick
skimming of its code; it seems to trigger whenever data is received on the
terminal's channel (regardless of whether it causes any text changes or not) and
also when the cursor moves, but only if it's visible.
Not clear to me how this event was actually intended to work...